add macro to impl ErrorMessaging
